Wrap the Cache in an UnsafeCell
This cache is only accessed by raw pointer through the config via the C lfs code.
This pointer however is used while the `RefCell` containing the allocation is borrowed,
therefore aliasing the borrow. Using an UnsafeCell here should allow the C code to have the only
mutable borrow on the underlying value.