Add Filesystem::mount_or_else
When mounting a filesystem, often code like this is used:

    if !Filesystem::is_mountable(alloc, storage) {
        Filesystem::format(storage).ok();
    }
    Filesystem::mount(alloc, storage)

This mounts the filesystem twice because Filesystem::is_mountable is
equivalent to Filesystem::mount(...).is_ok().  Depending on the storage
implementation, mounting the filesystem can have significant cost.
But directly calling Filesystem::mount and re-mounting in the error case
is prohibited by the borrow checker.

This patch adds a mount_or_else method that accepts a callable that is
called on mount error.  Afterwards, mounting is re-tried.