Remove cstr_core and cty dependencies and use littlefs2-sys 0.2.0
CStr, c_char, c_void and c_int are now available in core::ffi so we no
longer need dependencies for them.  c_size_t is currently only available
in nightly so we use a typedef instead.

As littlefs2-sys 0.1 is still using cty, we also need to update it to
0.2.0.