Replace Error enum with struct
To make it easier to add support for new error codes without breaking
compatibility, this patch replaces the Error enum with a struct and
associated constants.  As a part of this change, it also enforces that
Error is not used for return values indicating success, i. e. only for
negative return values.