Add DynStorageAlloc using FnOnce
Our DynStorage trait uses a &mut dyn FnMut as a callback for
implementing mount_and_then (instead of a FnOnce like Filesystem).  This
is because there is no no-std way to pass a FnOnce in an dyn compatible
trait method.

If alloc is available, this limitation can be avoided by using a
Box<dyn FnOnce>.  This patch adds an alloc feature to the crate and a
DynStorageAlloc trait behind that feature that provides mount_and_then
functions using a boxed FnOnce instead of a reference to an FnMut.