Remove the use of `RefCell` in `File`
The previous `RefCell` was only dealing with a pointer and the data was always
held by the lfs side so the checking from the refcell was useless anyways

This also contains a breaking change because the `unsafe` `open` function
did not take the correct lifetime.

It think it's an acceptable breaking change because the previous behaviour was buggy
and any code that depends on it likely has a use after free.